requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Traditional ML
scikit-learn>=1.3.0
xgboost>=2.0.0

# Explainability
grad-cam>=1.4.8
shap>=0.43.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Deployment
streamlit>=1.28.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
